pipeline/requirements.txt
# Core dependencies
llama-index-core>=0.10.0
llama-index-readers-boarddocs>=0.1.0
anthropic>=0.18.0
supabase>=2.0.0
resend>=0.7.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Development
pytest>=8.0.0
mypy>=1.8.0
